[fix / add] fix requirement.c and add flag 1 and flag 2

+++ lib/my/my_str_is_a_number.c
@@ -9,7 +9,7 @@
 
 bool my_str_is_num(char *str)
 {
-    long long int i = 0;
+    int i = 0;
 
     if (str == NULL)
         return false;
